chore: Add versions of Superpowered to project 0.9.7-1.0.3
Currently, 1.0.0 is used to build

+++ SuperpoweredSDK/Superpowered/SuperpoweredCPU.h
@@ -16,7 +16,7 @@
 */
 class SuperpoweredCPU {
 public:
-    static long sustainedPerformanceMode;
+    static int sustainedPerformanceMode;
 /**
  @brief Enable/disable the sustained performance mode. This method is thread-safe.
